# Migrate building CommonJS module entrypoint to vite