--- a/package.json
+++ b/package.json
@@ -31,11 +31,11 @@
     "cypress:gui": "TZ=UTC cypress open --component",
     "cypress:update-snapshots": "TZ=UTC cypress run --component --spec cypress/visual/**/*.cy.js --env type=base --config screenshotsFolder=cypress/snapshots/base"
   },
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
   "exports": {
     ".": {
       "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
+      "require": "./dist/index.cjs"
     },
     "./dist/": "./dist/"
   },
@@ -103,7 +103,7 @@
     "@nextcloud/eslint-config": "^8.3.0-beta.0",
     "@nextcloud/stylelint-config": "^2.3.1",
     "@nextcloud/webpack-vue-config": "github:nextcloud/webpack-vue-config#master",
-    "@susnux/nextcloud-vite-config": "^1.0.0-beta.13",
+    "@susnux/nextcloud-vite-config": "^1.0.0-beta.15",
     "@vue/test-utils": "^1.3.0",
     "@vue/tsconfig": "^0.4.0",
     "@vue/vue2-jest": "^29.0.0",