### ☑️ Resolves

* Follow up on #4276 

This uses vite also for the CommonJS module entry point (index).
* Much faster build time (18% faster)¹
* Much smaller output asset: 400kB (vite) vs 1'020kB (webpack)

I tested it with the *text* app, looks good, the js assets of the text app will shrink in total by 254kB with the vite version.

¹ Run time on my system (using `time npm run build`):
```
before:  npm run build  131,93s user 7,69s system 29,180 total
after:   npm run build  117,00s user 6,73s system 23,918 total
```
### 🏁 Checklist

- [x] ⛑️ Tests are included or are not applicable
- [x] 📘 Component documentation has been extended, updated or is not applicable